Rework and upgrade  of the extension frontend
This commit massively reworks the way the extension works on the
frontend:
- No more Odoo sidebar, Configurations navigation and modification is done through the
status bar widget
- Creating a new Configuration is now done through the webview instead
of the QuickPick menu
- Configurations are no longer stored in settings.json but are instead
hidden in the extension's state
- Refactoring of the client folder structure
- Display the Configuration's detected odoo version underneath the Odoo
folder path
- Ability to add multiple addons folders to a configuration
- It is now possible to have multiple Configurations webviews open at the
same time
- Autosave name and paths when focus is lost on the field or enter is
pressed
