+++ .vscodeignore
@@ -1,7 +1,6 @@
 .vscode
 .gitignore
 client/out/*.map
-client/src/
 tsconfig.json
 tslint.json
 package.json
