+++ client/utils/events.ts
@@ -1,0 +1,4 @@
+import {EventEmitter} from "vscode";
+
+export const selectedConfigurationChange = new EventEmitter();
+export const ConfigurationsChange = new EventEmitter();
